+++ Changelog.md
@@ -118,7 +118,7 @@
     1. Snapping on by default, off holding Shift. Improved snapping.
         * [Pull request #1304](https://bitbucket.org/osrf/gazebo/pull-request/1304)
 
-    1. Snap walls to length increments, moved scale to SegmentItem and added Get/SetScale.
+    1. Snap walls to length increments, moved scale to SegmentItem and added Get/SetScale, added SegmentItem::SnapAngle and SegmentItem::SnapLength.
         * [Pull request #1311](https://bitbucket.org/osrf/gazebo/pull-request/1311)
 
 1. Model editor updates